:art: [feat][products]: Helpers class created for keep helper methods.

+++ lib/core/helpers/helpers.dart
@@ -3,7 +3,7 @@ class Helpers{
   const Helpers._();
 
 
-  static double calculateDiscountedPrice(double amount, int discountPercent) {
+  static double calculateDiscountedPrice(double amount, double discountPercent) {
     if(amount <= 0 && discountPercent <= 0) return amount;
 
     final discount = (amount * discountPercent) / 100;
